Fixed debug printout

+++ NetworkModule/NetworkData.h
@@ -21,7 +21,7 @@ struct Packet
 {
 	PacketTypes packet_type;	// 4
 	int packet_ID;
-	int timestamp;
+	float timestamp;
 
 	virtual void serialize(char * data)
 	{			//Turn the PacketType into bytes
